fix - #11 로그인/아웃, RefreshToken 재발급 코드 구조 변경
Related to: #11
