+++ server/src/main/java/challenge/server/config/SecurityConfig.java
@@ -1,42 +1,24 @@
 package challenge.server.config;
 
-import challenge.server.security.jwt.JwtAuthenticationToken;
-import challenge.server.security.oauth.handler.OAuth2MemberSuccessHandler;
-import challenge.server.security.oauth.service.CustomOAuth2UserService;
-import challenge.server.security.filter.JwtAuthenticationFilter;
-import challenge.server.security.filter.JwtVerificationFilter;
 import challenge.server.security.handler.UserAccessDeniedHandler;
 import challenge.server.security.handler.UserAuthenticationEntryPoint;
-import challenge.server.security.handler.UserAuthenticationFailureHandler;
-import challenge.server.security.handler.UserAuthenticationSuccessHandler;
 import challenge.server.security.jwt.JwtTokenizer;
+import challenge.server.security.oauth.handler.OAuth2MemberSuccessHandler;
+import challenge.server.security.oauth.service.CustomOAuth2UserService;
 import challenge.server.security.utils.CustomAuthorityUtils;
 import challenge.server.user.service.LogoutListService;
 import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.config.ldap.LdapBindAuthenticationManagerFactory;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.util.StringUtils;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.CorsUtils;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.Arrays;
 
-import static org.springframework.http.HttpMethod.*;
 import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
@@ -64,8 +46,8 @@ public class SecurityConfig { // https 적용
                 .authenticationEntryPoint(new UserAuthenticationEntryPoint()) //Oauth2에서는 인증에서 실패했을때 처리하는 로직
                 .accessDeniedHandler(new UserAccessDeniedHandler()) //인가 에러 핸들링
                 .and()
-                .apply(new CustomFilterConfigurer())
-                .and()
+//                .apply(new CustomFilterConfigurer())
+//                .and()
                 // v1
                 .authorizeHttpRequests(authorize -> authorize.anyRequest().permitAll())
                 // v2
@@ -137,7 +119,7 @@ public class SecurityConfig { // https 적용
     /**
      * 삭제와 생성을 할수있도록 spring bean 등록해야함.
      * bean설정을 따로 해줘야한다. 안하면 service단에서 빈을 자동생성 못함.
-     *
+     * <p>
      * CORS 관련 설정
      */
     @Bean
@@ -155,24 +137,24 @@ public class SecurityConfig { // https 적용
         return source;
     }
 
-    public class CustomFilterConfigurer extends AbstractHttpConfigurer<CustomFilterConfigurer, HttpSecurity> {
-        @Override
-        public void configure(HttpSecurity builder) throws Exception {
-            AuthenticationManager authenticationManager = builder.getSharedObject(AuthenticationManager.class);
-//            builder.addFilter(new CustomFilter(authenticationManager));
-
-            JwtAuthenticationFilter jwtAuthenticationFilter = new JwtAuthenticationFilter(authenticationManager, jwtTokenizer, userService);
-            jwtAuthenticationFilter.setFilterProcessesUrl("/login");
-            jwtAuthenticationFilter.setAuthenticationSuccessHandler(new UserAuthenticationSuccessHandler());
-            jwtAuthenticationFilter.setAuthenticationFailureHandler(new UserAuthenticationFailureHandler());
-
-            JwtVerificationFilter jwtVerificationFilter = new JwtVerificationFilter(jwtTokenizer, authorityUtils, logoutListService);
-
-            builder
-                    .addFilter(jwtAuthenticationFilter)
-                    .addFilterAfter(jwtVerificationFilter, JwtAuthenticationFilter.class);
-        }
-    }
+//    public class CustomFilterConfigurer extends AbstractHttpConfigurer<CustomFilterConfigurer, HttpSecurity> {
+//        @Override
+//        public void configure(HttpSecurity builder) throws Exception {
+//            AuthenticationManager authenticationManager = builder.getSharedObject(AuthenticationManager.class);
+////            builder.addFilter(new CustomFilter(authenticationManager));
+//
+//            JwtAuthenticationFilter jwtAuthenticationFilter = new JwtAuthenticationFilter(authenticationManager, jwtTokenizer, userService);
+//            jwtAuthenticationFilter.setFilterProcessesUrl("/login");
+//            jwtAuthenticationFilter.setAuthenticationSuccessHandler(new UserAuthenticationSuccessHandler());
+//            jwtAuthenticationFilter.setAuthenticationFailureHandler(new UserAuthenticationFailureHandler());
+//
+//            JwtVerificationFilter jwtVerificationFilter = new JwtVerificationFilter(jwtTokenizer, authorityUtils, logoutListService);
+//
+//            builder
+//                    .addFilter(jwtAuthenticationFilter)
+//                    .addFilterAfter(jwtVerificationFilter, JwtAuthenticationFilter.class);
+//        }
+//    }
 
 //    @Bean
 //    public AuthenticationManager authenticationManagerBean() throws Exception {
