+++ server/src/main/java/challenge/server/security/filter/JwtVerificationFilter.java
@@ -1,30 +1,17 @@
 package challenge.server.security.filter;
 
-import challenge.server.exception.BusinessLogicException;
-import challenge.server.exception.ExceptionCode;
-import challenge.server.security.jwt.JwtAuthenticationToken;
 import challenge.server.security.jwt.JwtTokenizer;
 import challenge.server.security.utils.CustomAuthorityUtils;
 import challenge.server.user.entity.LogoutList;
 import challenge.server.user.service.LogoutListService;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.security.SignatureException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -34,8 +21,8 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.Map;
 
 @Component
 @Slf4j
@@ -76,26 +63,6 @@ public class JwtVerificationFilter extends OncePerRequestFilter {
         return authorization == null || !authorization.startsWith("Bearer");
     }
 
-    // JWT 토큰 복호화해서 토큰에 들어있는 정보 꺼냄
-    public Authentication getAuthentication(String accessToken, String base64EncodedSecretKey) {
-        // 토큰 복호화
-        Jws<Claims> claims = jwtTokenizer.getClaims(accessToken, base64EncodedSecretKey);
-
-        if (claims.getBody().getSubject() == null) { // 회원 email
-            throw new BusinessLogicException(ExceptionCode.AUTHENTICATION_NOT_FOUND);
-        }
-
-        // 클레임에서 권한 정보 가져오기
-        Collection<? extends GrantedAuthority> authorities =
-                Arrays.stream(claims.getBody().getSubject().toString().split(","))
-                        .map(SimpleGrantedAuthority::new)
-                        .collect(Collectors.toList());
-
-        // UserDetails 객체를 만들어서 Authentication 리턴
-        UserDetails principal = new User(claims.getBody().getSubject(), "", authorities);
-        return new UsernamePasswordAuthenticationToken(principal, "", authorities);
-    }
-
     private Map<String, Object> verifyJws(HttpServletRequest request) {
         String jws = request.getHeader("Authorization").replace("Bearer ", "");
         String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
@@ -118,52 +85,5 @@ public class JwtVerificationFilter extends OncePerRequestFilter {
         }
 
         return null;
-    }
-
-    private String delegateAccessToken(challenge.server.user.entity.User user) {
-        Map<String, Object> claims = new HashMap<>();
-//        claims.put("userId", user.getUserId());
-        claims.put("username", user.getUsername());
-        claims.put("roles", user.getRoles());
-
-        String subject = user.getEmail();
-        Date expiration = jwtTokenizer.getTokenExpiration(jwtTokenizer.getAccessTokenExpirationMinutes());
-        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-        String accessToken = jwtTokenizer.generateAccessToken(claims, subject, expiration, base64EncodedSecretKey);
-        return accessToken;
-    }
-
-    private String delegateRefreshToken(challenge.server.user.entity.User user) {
-        String subject = user.getEmail();
-        Date expiration = jwtTokenizer.getTokenExpiration((jwtTokenizer.getRefreshTokenExpirationMinutes()));
-        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
-        String refreshToken = jwtTokenizer.generateRefreshToken(subject, expiration, base64EncodedSecretKey);
-        return refreshToken;
-    }
-
-    public String reissueRefreshToken(challenge.server.user.entity.User user, HttpServletResponse response) throws ServletException, IOException {
-        String refreshToken = delegateRefreshToken(user);
-
-        response.setHeader("Refresh", refreshToken);
-
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpStatus.OK.value());
-
-//        Map<String, Long> body = new HashMap<>();
-//        body.put("userId", user.getUserId()); // 로그인 후 userId를 응답 body에 전달
-
-//        new ObjectMapper().writeValue(response.getOutputStream(), body);
-
-        return refreshToken;
-    }
-
-    public String reissueAccessToken(challenge.server.user.entity.User user, HttpServletResponse response) {
-        String accessToken = delegateAccessToken(user);
-        response.setHeader("Authorization", "Bearer " + accessToken);
-
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpStatus.OK.value());
-
-        return accessToken;
     }
 }
